split theme-park/index.js into constiuent component files
